requirements.txt
requests>=2.31.0
python-dotenv==1.0.0
yt-dlp==2023.7.6
librosa==0.10.1
numpy==1.24.3
ffmpeg-python>=0.2.0
spotipy==2.23.0
supabase==2.0.2
PyYAML==6.0.1
# Mozart enhanced analysis dependencies
pandas>=1.3.0
soundfile>=0.12.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
